feat: implement Universal Chain Support pilot and partner adapters (#142)1
Executed the Protocol Adapter Pilot and Partner lanes for Liquid (CON-710), Rootstock (CON-711), Babylon (CON-712), and BitVM (CON-713).

Key changes:
- Implemented ChainAdapter trait in conxian-core for unified multi-chain interaction.
- Created LiquidAdapter and RootstockAdapter pilot implementations in internal/engine.
- Added skeletal BabylonAdapter and BitVmAdapter for partner-lane research.
- Integrated multi-chain registry into AppState and exposed new API endpoints for chain height lookup and transaction preparation.
- Expanded research documentation with ADAPTER_FAMILY_STRATEGY.md, PARTNER_LANE_ADAPTERS.md, and VERIFICATION_IMPROVEMENT_PROPOSAL.md.
- Hardened library boundaries by standardizing Bitcoin amounts as u64 satoshis across the workspace.
- Updated TypeScript Client SDK with institutional-grade helper classes.
- Verified all changes with workspace tests, clippy, and formatting.

+++ internal/compliance/src/zkc.rs
@@ -527,7 +527,7 @@ impl ZkcVerifier {
     pub fn sign_offline_receipt(
         &self,
         tx_hash: &str,
-        amount_sbtc: f64,
+        amount_sbtc: u64,
         device_id: &str,
         passkey_attestation: AttestationRequest,
     ) -> ConxianResult<OfflineReceipt> {
